+++ lib/data/index.ts
@@ -1,3 +1,6 @@
 export * from "./counties";
 export * from "./constituencies";
 export * from "./wards";
+export * from "./sub-counties";
+export * from "./locality";
+export * from "./area";